Add files via upload
INX Future Inc, is one of the leading data analytics and automation solutions provider with over 15 years of global business presence. In recent years, the employee performance indexes are not healthy and this has become a growing concern among the top management. The CEO Mr. Brain, decided to initiate a data science project, which analyzes the current employee data and find the core underlying causes of the performance issues. He also expects a clear indicators of non-performing employees, so that any penalization of non-performing employee, if required, may not significantly affect other employee morals. The following insights are expected from this project:

Department wise performances.
Top-3 Important Factors effecting employee performance.
A trained model which can predict the employee performance based on factors as inputs.
Recommendations to improve the employee performance based on insights from analysis.