Add files via upload
FIFA 20 is a football simulation video game published by Electronic Arts (EA) as part of the FIFA series. It is the 27th installment in the series. 
The aim of the project was to create a clustering model for categorizing or clustering the football players on the basis of their field positions and skillsets like shooting, passing, dribbling, goalkeeping, etc.