Add files via upload
An existential problem for any loan provider today is to find out the loan applicants who are very likely to repay the loan. By doing this, loan distributing institutions or banks can avoid huge losses and can work with less NPA (Non-Performing Assets).
The project goal was to predict whether the applicant/ bank customer will be able to repay the loan on time or be going to default the loan on the basis of the customer’s financial history with the banks.