Add files via upload
FicZon Inc is an IT solution provider with products ranging from on-premises products to SaaS-based solutions. FicZon’s business is majorly dependent on the sales force effectiveness. 
Effective sales are dependent on lead quality and as of now, this is based on manual categorization and highly dependent on the sales staff. 
The aim of the project was to explore Machine Learning to pre-categorize the lead quality and as result, expect a significant increase in sales effectiveness.